requirements.txt
apyori==1.1.1
beautifulsoup4==4.6.0
boto3==1.7.62
botocore==1.10.62
dj-database-url==0.5.0
Django>=2.1.5
django-daterange-filter==1.3.0
django-forms-bootstrap==3.1.0
django-materialize-form==0.1.1
django-materializecss-form==1.1.11
django-rest-framework==0.1.0
django-storages==1.6.6
djangorestframework==3.9.2
gunicorn==19.9.0
lxml==4.2.3
nltk==3.3
numpy==1.14.5
pandas==0.23.3
parso==0.2.1
pexpect==4.6.0
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2==2.7.7
ptyprocess==0.6.0
Pygments==2.2.0
requests==2.20.0
s3transfer==0.1.13
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
urllib3==1.23
